Initial swift argument parser documentation
