+++ Package.swift
@@ -11,9 +11,18 @@ let package = Package(
         .executable(name: pkgName, targets: [pkgName]),
         .library(name: libTarget, targets: [libTarget]),
     ],
-    dependencies: [ .package(url: "https://github.com/rhx/SwiftLibXML.git", .branch("main")) ],
+    dependencies: [ 
+        .package(url: "https://github.com/rhx/SwiftLibXML.git", .branch("main")),
+        .package(url: "https://github.com/apple/swift-argument-parser", from: "0.4.0"),
+    ],
     targets: [
-        .target(name: pkgName, dependencies: [.init(stringLiteral: libTarget)]),
+        .target(
+            name: pkgName, 
+            dependencies: [
+                .init(stringLiteral: libTarget),
+                .product(name: "ArgumentParser", package: "swift-argument-parser")
+            ]
+        ),
         .target(name: libTarget, dependencies: ["SwiftLibXML"]),
         .testTarget(name: "\(pkgName)Tests", dependencies: [.init(stringLiteral: libTarget)]),
     ]
